School-Web_design
The project is the realization of a multi-page site on the topic of Šokci and their costumes. The site contains the realized description and explanation of certain terms related to the topic, gallery, Site map and RSS. This project required learning HTML and CSS, as well as the distribution of content by pages. Database doesn’t exist.
Technologies – HTML, CSS